archive/scripts/requirements.txt
IMDbPY==2021.4.18
PyYAML==5.4.1
pyperclip==1.8.2
python-frontmatter==1.0.0
